+++ src/pyquery_polars/__init__.py
@@ -2,4 +2,4 @@ from .backend.engine import PyQueryEngine
 
 __all__ = ["PyQueryEngine"]
 
-__version__ = "1.3.0"
+__version__ = "2.3.0"